Python_ICANH_Project/requirements.txt
fastapi==0.112.4
uvicorn[standard]==0.30.6
sqlalchemy==2.0.31
alembic==1.13.2
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
